+++ examples/resources/project_member/import.sh
@@ -1,1 +1,0 @@
-terraform import project_member.mymember project_key:username