+++ examples/resources/project_user/resource.tf
@@ -1,4 +1,4 @@
-resource "project_member" "myuser" {
+resource "project_user" "myuser" {
   project_key = "myproj"
   name        = "myuser"
   roles       = ["Viewer"]