+++ pkg/project/provider.go
@@ -51,7 +51,8 @@ func Provider() *schema.Provider {
 				"project":             projectResource(),
 				"project_environment": projectEnvironmentResource(),
 				"project_role":        projectRoleResource(),
-				"project_member":      projectMemberResource(),
+				"project_user":        projectUserResource(),
+				"project_group":       projectGroupResource(),
 			},
 		),
 	}